Changes naming convention and Adds new functio to main
getTableBody was renamed to getTable for more accuracy in describing
the behaviour of the function.

WipeTableBody() successfully selects the table body element and deletes
it. The function clears the body so that the matches obtained by
searching via the search box can be appended to a fresh table,
successfully filtering out the rest.